backend/requirements/local.txt
-r base.txt

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
isort==5.13.2

# Development
ipython==8.19.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
